# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/car": {
            "get": {
                "tags": [
                    "Carros"
                ],
                "summary": "Buscar todas as informações",
                "operationId": "ad16506d8ae145d931d69365fdd67765",
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Carros"
                ],
                "summary": "Criar novo carro",
                "operationId": "f8c3a41fbed6397a429d561fd82e0b62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/car/{parking_id}/{car_id}": {
            "get": {
                "tags": [
                    "Carros"
                ],
                "summary": "Buscar carro por ID do estacionamento e ID do carro",
                "operationId": "63b8a41fa2907b5d27063bcf68b698f2",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Carros"
                ],
                "summary": "Deletar carro por ID do estacionamento e ID do funcionário",
                "operationId": "f0f8d3d3ea06554e03a0698a77ae7c36",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Carros"
                ],
                "summary": "Atualizar carro por ID do estacionamento e ID do carro",
                "operationId": "4a69a074addcb17150091b61e9e53aa0",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/car/output/{parking_id}/{car_id}": {
            "patch": {
                "tags": [
                    "Carros"
                ],
                "summary": "Adicionar saida para por ID do estacionamento e ID do carro",
                "operationId": "8fe341668cea40f2f91ce859cfba5f97",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarsDTO": {
                "properties": {
                    "plate": {
                        "type": "string",
                        "example": "ABC-123"
                    },
                    "model": {
                        "type": "string",
                        "example": "Uno"
                    },
                    "color": {
                        "type": "string",
                        "example": "Verde"
                    },
                    "parking_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "CarResource": {
                "properties": {
                    "errors": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "content": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "placa": {
                                "type": "string"
                            },
                            "modelo": {
                                "type": "strign"
                            },
                            "cor": {
                                "type": "string"
                            },
                            "entrada": {
                                "type": "string",
                                "format": "date"
                            },
                            "estacionamento_id": {
                                "type": "integer"
                            },
                            "saida": {
                                "type": "string",
                                "format": "date"
                            },
                            "valor_para_pagamento": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}